refactor(content): WKC layout — Phase 9 (controllers folder rename)
Phase 9 of the WKC migration. Renames src/controller/ to src/controllers/
to match the WKC convention (plural matches the per-handler-file
contents and aligns with the standard layout documented in CLAUDE.md).

The handlers themselves were already thin (validate input, call logic
component, return response) since Phase 1 set up the convention. No
business logic was leaking into the handlers, so no audit-and-extract
changes were required for this PR.

Wiring:

- All importer paths swept via sed (from '.../controller/' or './controller/'
  to '.../controllers/' / './controllers/').
- Folder renamed via git mv to preserve history.

tsc clean, unit suite green (213 tests, no diff).
