+++ content/src/controllers/handlers/get-available-content-handler.ts
