+++ content/src/controllers/handlers/get-challenge-handler.ts
