+++ content/src/controllers/handlers/get-entities-handler.ts
