+++ content/src/controllers/handlers/get-entity-image-handler.ts
