+++ content/src/controllers/handlers/get-entity-thumbnail-handler.ts
