+++ content/src/controllers/handlers/get-erc721-entity-handler.ts
