+++ content/src/controllers/handlers/get-snapshots-handler.ts
