+++ content/test/unit/controllers/conditional-request.spec.ts
@@ -1,9 +1,9 @@
 import { ContentItem, IContentStorageComponent } from '@dcl/catalyst-storage'
 import { HandlerContextWithPath } from '../../../src/types'
-import { getContentHandler } from '../../../src/controller/handlers/get-content-handler'
-import { getEntityImageHandler } from '../../../src/controller/handlers/get-entity-image-handler'
-import { getEntityThumbnailHandler } from '../../../src/controller/handlers/get-entity-thumbnail-handler'
-import { toETag } from '../../../src/controller/utils'
+import { getContentHandler } from '../../../src/controllers/handlers/get-content-handler'
+import { getEntityImageHandler } from '../../../src/controllers/handlers/get-entity-image-handler'
+import { getEntityThumbnailHandler } from '../../../src/controllers/handlers/get-entity-thumbnail-handler'
+import { toETag } from '../../../src/controllers/utils'
 import { createContentItemMock } from '../../mocks/content-item-mock'
 import { createStorageComponentMock } from '../../mocks/storage-component-mock'
 import { createRequestMock } from '../../mocks/request-mock'
